mainapp/requirements.txt
django>=3.0.5
djangorestframework>=3.11.0,<3.12.0
wheel>=0.34.2
psycopg2-binary
dj-database-url>=0.5.0
uwsgi>=2.0.18,<2.1.0